Stabilized integrator
The integration was unstable because the community was subject to invasion by species not present in the initial conditions. Removing these species during the integration was somewhat challenging, given the existing architecture, but now everything is very stable.

The central change was exposing the parameter values to IntegrateWell, so that the matrices could be compressed. This required two modifications in dydt, one to add the parameters as an additional argument, and the other to allow the dividing line between species and resources to be modified.