Fixed problems with plotting timeseries
Main problem was that the current state pointed to the same object as
the initial state whenever the initial state was a dataframe. This
meant that if you initialized with a dataframe, ran a simulation, and
then initialized a new community with the same dataframe, the two
initial states would be different. I was safe in model_calc.py because
I passed N0.values instead of N0 to run_model.